// frontend/src/app/api/generated/model/predictionResponse.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UpcomingPayment } from './upcomingPayment';
import { MonthlyPrediction } from './monthlyPrediction';


export interface PredictionResponse { 
    predictions: Array<MonthlyPrediction>;
    upcomingPayments: Array<UpcomingPayment>;
}